requirements.txt
attrs==23.2.0
Automat==22.10.0
beautifulsoup4==4.12.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
constantly==23.10.4
cryptography==42.0.2
cssselect==1.2.0
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.13.1
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
itemadapter==0.8.0
itemloaders==1.1.0
jieba3k==0.35.1
jmespath==1.0.1
joblib==1.3.2
lxml==5.1.0
newspaper3k==0.2.8
nltk==3.8.1
packaging==23.2
parsel==1.8.1
pillow==10.2.0
Protego==0.3.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyDispatcher==2.0.7
pyOpenSSL==24.0.0
python-dateutil==2.8.2
PyYAML==6.0.1
queuelib==1.6.2
random-user-agent==1.0.1
regex==2023.12.25
requests==2.31.0
requests-file==2.0.0
Scrapy==2.11.0
service-identity==24.1.0
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.5
tinysegmenter==0.3
tldextract==5.1.1
tqdm==4.66.1
Twisted==22.10.0
typing_extensions==4.9.0
urllib3==2.2.0
w3lib==2.1.2
zope.interface==6.1
kafka~=1.3.5
pyspark~=3.5.0
schedule~=1.2.1
cassandra-driver~=3.29.0